==> 1) filling nan values train_df.ipynb <==
import numpy as np
import pandas as pd

from sklearn.linear_model import LogisticRegression

#ignoring the warnings
import sys
if not sys.warnoptions:
    import warnings
    warnings.simplefilter("ignore")

def clean(df):
    df['income']=df['ApplicantIncome']+df['CoapplicantIncome']
    df['loanpm']=df['LoanAmount']+df['Loan_Amount_Term']
    df.drop(['LoanAmount','Loan_ID','Loan_Amount_Term','ApplicantIncome','CoapplicantIncome'],axis='columns',inplace=True)
    df['Gender']=df['Gender'].replace({'Male':1,'Female':0})
    df['Married']=df['Married'].replace({'Yes':1,'No':0})
    df['Education']=df['Education'].replace({'Graduate':1,'Not Graduate':0})
    df['Self_Employed']=df['Self_Employed'].replace({'Yes':1,'No':0})
    df['Dependents']=df['Dependents'].replace({'3+':3,'1':1,'2':2})
    df['Property_Area']=df['Property_Area'].replace({'Urban':3,'Rural':1,'Semiurban':2})
    df['income_bins']=pd.cut(df.income,np.linspace(1400,81000,dtype='int',num=20),labels=np.linspace(0,18,dtype='int',num=19))
    df['Income']=pd.cut(df.income_bins,[-1,0,1,2,3,18],labels=[0,1,2,3,4])
    df['loan']=pd.cut(df.loanpm,np.linspace(100,1000,dtype='int',num=10),labels=[0,1,2,3,4,5,6,7,8])
    df.drop(['income','loanpm','income_bins'],axis=1,inplace=True)
    df=df.dropna()
    df.Dependents=df.Dependents.astype('int')
    return df

#filling the unknown values of columns rather than droping
def get_filled(train1,col):
    train1=train.loc[train[col].isnull()]
    train1[col]='0'
    print(train1.isnull().sum())
    return train1

def new_df(df1,df2,col,df3):
    df_new=pd.concat([df1,df2],axis=0)
    df_new.drop('Loan_Status',axis=1,inplace=True)
    df3.drop('Loan_Status',axis=1,inplace=True)
    df_new=clean(df_new)
    x=df_new.drop(col,axis=1)
    y=df_new[col]
    y=y.astype('int')
    y.columns=[col]
    lreg=LogisticRegression()
    lreg.fit(x,y)
    df3[col]=lreg.predict(df3.drop(col,axis=1))
    a=df2.loc[df2[col].notnull(),col]
    b=df3[col]
    c=pd.concat([a,b],axis=0)
    c.sort_index(axis=0,ascending=True,inplace=True)
    df2.drop(col,axis=1,inplace=True)
    df2=pd.concat([df2,c],axis=1)
    return df2
    #df1==df(cleaned)
    #df2==train(cleaned wrt col)
    #df3==train(cleaned)

df=pd.read_csv('test.csv')
df.head()

train0=pd.read_csv('train.csv')
train=train0.copy()
train.head()

train.isnull().sum()

# filling the unknown values of gender

train1=get_filled(train,'Gender')

train1.Credit_History.replace({np.nan:0},inplace=True)
train1=clean(train1)
train1.isnull().sum()

train['Gender']=train['Gender'].replace({'Male':'1','Female':'0'})
train=new_df(df,train,'Gender',train1)
train.isnull().sum()

# filling the unknown values of dependents

train1=get_filled(train,'Dependents')

train1.Loan_Amount_Term.median()

train1.LoanAmount.median()

train1.Married.mode()

train1.Loan_Amount_Term.replace({np.nan:360},inplace=True)
train1.LoanAmount.replace({np.nan:138},inplace=True)
train1.Married.replace({np.nan:'Yes'},inplace=True)
train1['Gender']=train1['Gender'].replace({1:'Male',0:'Female'})

train1=clean(train1)
train1.isnull().sum()

train['Dependents']=train['Dependents'].replace({'3+':3,'1':1,'2':2})
train=new_df(df,train,'Dependents',train1)
train.isnull().sum()

# filling the unknown values of self_employed

train1=get_filled(train,'Self_Employed')

train.Credit_History.mode()

train1.LoanAmount.median()

train1.Loan_Amount_Term.median()

train1.Credit_History.replace({np.nan:1},inplace=True)
train1.Loan_Amount_Term.replace({np.nan:360},inplace=True)
train1.LoanAmount.replace({np.nan:125},inplace=True)
train1['Gender']=train1['Gender'].replace({1:'Male',0:'Female'})
train1['Dependents']=train1['Dependents'].replace({'3+':3,'1':1,'2':2})
train1=clean(train1)
train1.isnull().sum()

train['Self_Employed']=train['Self_Employed'].replace({'Yes':1,'No':0})
train=new_df(df,train,'Self_Employed',train1)
train.isnull().sum()

# filling the unknown values of credit_history

train1=get_filled(train,'Credit_History')

train1.LoanAmount.median()

train1.LoanAmount.replace({np.nan:130},inplace=True)
train1['Gender']=train1['Gender'].replace({1:'Male',0:'Female'})
train1['Dependents']=train1['Dependents'].replace({'3+':3,'1':1,'2':2})
train1['Self_Employed']=train1['Self_Employed'].replace({1:'Yes',0:'No'})
train1=clean(train1)
train1.isnull().sum()

train=new_df(df,train,'Credit_History',train1)
train.isnull().sum()

train.Credit_History.fillna(1,inplace=True)
train.isnull().sum()

train.LoanAmount.median()

train.Loan_Amount_Term.median()

train.LoanAmount.replace({np.nan:125},inplace=True)
train.Loan_Amount_Term.replace({np.nan:360},inplace=True)
train.isnull().sum()

# filling the unknown values of married

train1=get_filled(train,'Married')

train1['Gender']=train1['Gender'].replace({1:'Male',0:'Female'})
train1['Dependents']=train1['Dependents'].replace({3:'3+',1:'1',2:'2'})
train1['Dependents']=train1['Dependents'].astype('object')
train1.dtypes

train1['Self_Employed']=train1['Self_Employed'].replace({1:'Yes',0:'No'})
train1=clean(train1)
train1.isnull().sum()

train=new_df(df,train,'Married',train1)
train.isnull().sum()

train.head()

train['Gender']=train['Gender'].replace({1:'Male',0:'Female'})
train['Dependents']=train['Dependents'].replace({3:'3+',1:'1',2:'2'})
train['Self_Employed']=train['Self_Employed'].replace({1:'Yes',0:'No'})
train=clean(train)

train.head()

train.Loan_Status=train0.Loan_Status

train=train[['Gender','Married','Dependents','Education','Self_Employed','Credit_History','Property_Area','Income','loan','Loan_Status']]
train.head()

train.to_csv('train_modified.csv',index=False)
#saving training data set
