==> 3) model selection.ipynb <==
import numpy as np
import pandas as pd

df=pd.read_csv('train_modified.csv')
df.head()

x=df.drop('Loan_Status',axis=1)
y=df['Loan_Status']

from sklearn.model_selection import cross_val_score
from sklearn.model_selection import KFold
scoring='accuracy'
kf=KFold(n_splits=10,random_state=1,shuffle=True)

from sklearn.tree import DecisionTreeClassifier
list=[5,10,50,100,500]
for i in list:
    des=DecisionTreeClassifier(max_leaf_nodes=i,random_state=1)
    score=cross_val_score(des,x,y,cv=kf,scoring=scoring).mean()
    print("..",i)
    print(score)

from sklearn.tree import DecisionTreeClassifier
list=[2,3,4,5,6,7,8,9]
for i in list:
    des=DecisionTreeClassifier(max_leaf_nodes=i,random_state=1)
    score=cross_val_score(des,x,y,cv=kf,scoring=scoring).mean()
    print("..",i)
    print(score)

des=DecisionTreeClassifier(max_leaf_nodes=2,random_state=1)
score=cross_val_score(des,x,y,cv=kf,scoring=scoring).mean()
score

# Decisiontree -- 80.92

from sklearn.ensemble import RandomForestClassifier
list=[5,10,50,100,500]
for i in list:
    rf=RandomForestClassifier(max_leaf_nodes=i,random_state=1)
    score=cross_val_score(rf,x,y,cv=kf,scoring=scoring).mean()
    print("..",i)
    print(score)

from sklearn.ensemble import RandomForestClassifier
list=[6,7,8,9,10,15,20,30,50]
for i in list:
    rf=RandomForestClassifier(max_leaf_nodes=i,random_state=1)
    score=cross_val_score(rf,x,y,cv=kf,scoring=scoring).mean()
    print("..",i)
    print(score)

rf=RandomForestClassifier(max_leaf_nodes=8,random_state=1)
score=cross_val_score(rf,x,y,cv=kf,scoring=scoring).mean()
score

# Randomforest -- 81.09 

from sklearn.linear_model import LogisticRegression
lreg=LogisticRegression()
score=cross_val_score(lreg,x,y,cv=kf,scoring=scoring).mean()
score

# Logistic reg -- 80.92

from sklearn.svm import SVC
svc=SVC()
score=cross_val_score(svc,x,y,cv=kf,scoring=scoring).mean()
score

# SVC -- 80.60

from sklearn.naive_bayes import GaussianNB
nb=GaussianNB()
score=cross_val_score(nb,x,y,cv=kf,scoring=scoring).mean()
score

# Navie bayes -- 80.76

from sklearn.neighbors import KNeighborsClassifier
knn=KNeighborsClassifier()
score=cross_val_score(knn,x,y,cv=kf,scoring=scoring).mean()
score

accuracy_rate=[]
for i in range(1,40):
    knn=KNeighborsClassifier(n_neighbors=i)
    score=cross_val_score(knn,x,y,cv=kf,scoring=scoring).mean()
    accuracy_rate.append(score)
np.max(accuracy_rate)

# KNN -- 74.71

# thus we decide to use regression forest classifier as our model
