==> 4) filling nan values test_df.ipynb <==
import numpy as np
import pandas as pd
from sklearn.linear_model import LogisticRegression

import sys
if not sys.warnoptions:
    import warnings
    warnings.simplefilter("ignore")

def clean(df):
    df['income']=df['ApplicantIncome']+df['CoapplicantIncome']
    df['loanpm']=df['LoanAmount']+df['Loan_Amount_Term']
    df.drop(['LoanAmount','Loan_ID','Loan_Amount_Term','ApplicantIncome','CoapplicantIncome'],axis='columns',inplace=True)
    df['Gender']=df['Gender'].replace({'Male':1,'Female':0})
    df['Married']=df['Married'].replace({'Yes':1,'No':0})
    df['Education']=df['Education'].replace({'Graduate':1,'Not Graduate':0})
    df['Self_Employed']=df['Self_Employed'].replace({'Yes':1,'No':0})
    df['Dependents']=df['Dependents'].replace({'3+':3,'1':1,'2':2})
    df['Property_Area']=df['Property_Area'].replace({'Urban':3,'Rural':1,'Semiurban':2})
    df['income_bins']=pd.cut(df.income,np.linspace(1400,81000,dtype='int',num=20),labels=np.linspace(0,18,dtype='int',num=19))
    df['Income']=pd.cut(df.income_bins,[-1,0,1,2,3,18],labels=[0,1,2,3,4])
    df['loan']=pd.cut(df.loanpm,np.linspace(100,1000,dtype='int',num=10),labels=[0,1,2,3,4,5,6,7,8])
    df.drop(['income','loanpm','income_bins'],axis=1,inplace=True)
    df=df.dropna()
    df.Dependents=df.Dependents.astype('int')
    return df

def get_filled(test1,col):
    test1=test.loc[test[col].isnull()]
    test1[col]='0'
    print(test1.isnull().sum())
    return test1

def new_df(df1,df2,col,df3):
    df_new=pd.concat([df1,df2],axis=0)
    df_new.drop('Loan_Status',axis=1,inplace=True)
    df_new=clean(df_new)
    x=df_new.drop([col],axis=1)
    y=df_new[col]
    lreg=LogisticRegression()
    lreg.fit(x,y)
    df3[col]=lreg.predict(df3.drop([col],axis=1))
    a=df2.loc[df2[col].notnull(),col]
    b=df3[col]
    c=pd.concat([a,b],axis=0)
    c.sort_index(axis=0,ascending=True,inplace=True)
    df2.drop(col,axis=1,inplace=True)
    df2=pd.concat([df2,c],axis=1)
    return df2
    #df1==df(cleaned)
    #df2==test(cleaned wrt col)
    #df3==test1(cleaned)

df=pd.read_csv('train.csv')
df.head()

test0=pd.read_csv('test.csv')
test=test0.copy()
test0.head()

test.isnull().sum()

test1=get_filled(test,'Gender')

test1.Dependents.replace({np.nan:0},inplace=True)
test1=clean(test1)
test1.isnull().sum()

test['Gender']=test['Gender'].replace({'Male':1,'Female':0})
test=new_df(df,test,'Gender',test1)
test.isnull().sum()

# gender done

test1=get_filled(test,'Dependents')

test1.Credit_History.mode()

test1.LoanAmount.median()

test1.Credit_History.replace({np.nan:1},inplace=True)
test1.LoanAmount.replace({np.nan:104},inplace=True)
test1['Gender']=test1['Gender'].replace({1:'Male',0:'Female'})

test1=clean(test1)
test1.isnull().sum()

test['Dependents']=test['Dependents'].replace({'3+':3,'1':1,'2':2})
test=new_df(df,test,'Dependents',test1)
test.isnull().sum()

# dependents done

test1=get_filled(test,'Self_Employed')

test1.Credit_History.replace({np.nan:1},inplace=True)
test1['Gender']=test1['Gender'].replace({1:'Male',0:'Female'})
test1['Dependents']=test1['Dependents'].replace({'3+':3,'1':1,'2':2})
test1=clean(test1)
test1.isnull().sum()

test['Self_Employed']=test['Self_Employed'].replace({'Yes':1,'No':0})
test=new_df(df,test,'Self_Employed',test1)
test.isnull().sum()

# self employed done

test1=get_filled(test,'Credit_History')

test1.Loan_Amount_Term.median()

test1.Loan_Amount_Term.replace({np.nan:360},inplace=True)
test1['Gender']=test1['Gender'].replace({1:'Male',0:'Female'})
test1['Dependents']=test1['Dependents'].replace({'3+':3,'1':1,'2':2})
test1['Self_Employed']=test1['Self_Employed'].replace({1:'Yes',0:'No'})
test1=clean(test1)
test1.isnull().sum()

test=new_df(df,test,'Credit_History',test1)
test.isnull().sum()

# credit history done

test.LoanAmount.median()

test.Loan_Amount_Term.median()

test.LoanAmount.replace({np.nan:125},inplace=True)
test.Loan_Amount_Term.replace({np.nan:360},inplace=True)
test.isnull().sum()

test.head()

test['Gender']=test['Gender'].replace({1:'Male',0:'Female'})
test['Dependents']=test['Dependents'].replace({3:'3+',1:'1',2:'2'})
test['Self_Employed']=test['Self_Employed'].replace({1:'Yes',0:'No'})
test=clean(test)

test.head()

test=test[['Gender','Married','Dependents','Education','Self_Employed','Credit_History','Property_Area','Income','loan']]
test.head()

test.to_csv('test_modified.csv',index=False)
