==> 5) loan predicting (RegressionForestClassifier).ipynb <==
import numpy as np
import pandas as pd

df=pd.read_csv('train_modified.csv')
df.head()

test=pd.read_csv('test_modified.csv')
test.head()

x=df.drop('Loan_Status',axis=1)
y=df.Loan_Status

submit=pd.read_csv(r'C:\Users\mansi\Desktop\data_science\loanpred\sample.csv')
submit.head()

from sklearn.ensemble import RandomForestClassifier
rf=RandomForestClassifier(max_leaf_nodes=8,random_state=1)
rf.fit(x,y)
submit['Loan_Status']=rf.predict(test)

submit.Loan_Status.value_counts()

y.value_counts()

submit.to_csv('final.csv',index=False)
